move application logic to new module
