+++ src/public-modules/Bounty/sagas.js
@@ -33,8 +33,7 @@ const {
   EXTEND_DEADLINE,
   INCREASE_PAYOUT,
   CONTRIBUTE,
-  TRANSFER_OWNERSHIP,
-  CREATE_FULFILLER_APPLICATION
+  TRANSFER_OWNERSHIP
 } = actionTypes;
 const {
   getBountySuccess,
@@ -44,9 +43,7 @@ const {
   stdBountySuccess,
   stdBountyFail,
   getDraftSuccess,
-  getDraftFail,
-  createFulfillerApplicationSuccess,
-  createFulfillerApplicationFail
+  getDraftFail
 } = actions;
 
 const {
@@ -589,24 +586,6 @@ export function* transferIssuer(action) {
   }
 }
 
-export function* createFulfillerApplication(action) {
-  const { bountyId, message, callback = () => {} } = action;
-
-  let endpoint = `bounty/${bountyId}/application/`;
-  let methodType = 'POST';
-
-  try {
-    const response = yield call(request, endpoint, methodType, {
-      data: { message }
-    });
-    yield put(createFulfillerApplicationSuccess());
-  } catch (e) {
-    yield put(createFulfillerApplicationFail(e));
-  }
-
-  callback();
-}
-
 export function* watchCreateDraft() {
   yield takeLatest([CREATE_DRAFT, UPDATE_DRAFT], createOrUpdateDraft);
 }
@@ -647,10 +626,6 @@ export function* watchContribute() {
   yield takeLatest(CONTRIBUTE, contribute);
 }
 
-export function* watchCreateFulfillerApplication() {
-  yield takeLatest(CREATE_FULFILLER_APPLICATION, createFulfillerApplication);
-}
-
 export default [
   watchGetDraft,
   watchCreateDraft,
@@ -661,6 +636,5 @@ export default [
   watchExtendDeadline,
   watchIncreasePayout,
   watchTransferIssuer,
-  watchContribute,
-  watchCreateFulfillerApplication
+  watchContribute
 ];
