+++ src/public-modules/FulfillerApplication/selectors.js
@@ -1,0 +1,1 @@
+import { createSelector } from 'reselect';
